% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad.R
\name{bad}
\alias{bad}
\title{badly documented function}
\usage{
bad(x)
}
\arguments{
\item{x}{input}

\item{...}{I factored this out and forgot to delete the Roxygen entry}
}
\description{
badly documented function
}
